Add depth and FindSubPath to FileTreeNode
A new property 'Depth' and a new method 'FindSubPath' were added to 'FileTreeNode'. 'Depth' is a ushort that represents a node's depth in the tree and can be used for various calculations. 'FindSubPath' is a function that returns all nodes that match a given subpath. This enhancement will allow more complex manipulations and considerations of the file tree hierarchy. Meanwhile, a test for 'FindSubPath' has been implemented in 'RelativeFileTreeTests'.