Make RelativePath constructor internal
https://github.com/Nexus-Mods/NexusMods.App/pull/2905 has showed that a
public constructor of `RelativePath` can result in unforseen issues. The
entire point of `RelativePath` and `AbsolutePath` is to provide a
consistent abstraction over file paths. To achieve this consistency,
paths have to be sanitized, otherwise many assumptions made in code
aren't valid anymore.

Safety goes out the window with a public constructor that doesn't
sanitized. I've added a `CreateUnsafe` method for the few cases where
you want to skip sanitization because the input is already sanitized.